+++ CHANGELOG.md
@@ -7,13 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-### [1.0.0] - 2023-01-31
+### [0.2.0] - 2023-02-02
+
+#### Added
+- Upgrade command, to automatically manage upgrades of all versioned files.
+- Check command, which will validate the changelog file. Optionally can trigger a system exit.
+- Content command, which will display the changelog content for a specific version.
+
+### [0.1.0] - 2023-01-31
 
 #### Changed
 
-- Updated "Tested up to" from 6.0.1 => 6.1.1
+- Initial Workings of the Changelog file.
 
 ### [LINKS]
 
-[Unreleased]: https://github.com/award28/changelogger/compare/1.0.0...HEAD
-[1.0.0]: https://github.com/award28/changelogger/commit/a284a424df9fe41121f5101b9a9471aa91fad7c4
+[Unreleased]: https://github.com/award28/changelogger/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/award28/changelogger/compare/0.1.0...0.2.0
+[0.1.0]: https://github.com/award28/changelogger/commit/a284a424df9fe41121f5101b9a9471aa91fad7c4